LLM/scripts/requirements.txt
# System dependencies required (install with apt):
# sudo apt-get update
# sudo apt-get install -y python3-dev python3-setuptools python3-pip python3-wheel
# sudo apt-get install -y build-essential libssl-dev libffi-dev
# sudo apt-get install -y espeak-ng python3-espeak
# sudo apt-get install -y ffmpeg

# CUDA/cuDNN dependencies (install in order):
# 1. Download cuDNN:
# wget https://developer.download.nvidia.com/compute/cudnn/9.10.1/local_installers/cudnn-local-repo-ubuntu2204-9.10.1_1.0-1_amd64.deb
# 2. Install cuDNN:
# sudo dpkg -i cudnn-local-repo-ubuntu2204-9.10.1_1.0-1_amd64.deb
# sudo cp /var/cudnn-local-repo-ubuntu2204-9.10.1/cudnn-*-keyring.gpg /usr/share/keyrings/
# sudo apt-get update
# sudo apt-get -y install cudnn
# sudo apt-get -y install cudnn-cuda-12
# sudo apt-get install libcudnn9-cuda-12
# sudo apt-get install libcudnn9-dev-cuda-12

# Python packages (install in this exact order):
numpy==1.22.0
torch --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118
scikit-learn==1.2.2
flask==2.0.1
Werkzeug==2.0.3
faster-whisper==0.9.0
transformers==4.30.0
soundfile==0.12.1
librosa==0.10.1
scipy==1.11.3
numba==0.56.4
llvmlite==0.39.1
ffmpeg-python==0.2.0

# Install TTS from GitHub for latest stable version
git+https://github.com/coqui-ai/TTS

# Optional dependencies
gradio
plotly
pandas
pycoingecko